updated DB context to work with domain models instead of dtos; cleanup;

+++ src/Eudaimonia.Domain/Rating.cs
@@ -2,7 +2,7 @@
 
 namespace Eudaimonia.Domain;
 
-public class Rating : Enumeration<int>
+public class Rating : Enumeration<byte>
 {
     public static readonly Rating OneStar = new("1", 1);
     public static readonly Rating TwoStar = new("2", 2);
@@ -10,7 +10,7 @@ public class Rating : Enumeration<int>
     public static readonly Rating FourStar = new("4", 4);
     public static readonly Rating FiveStar = new("5", 5);
 
-    private Rating(string name, int value) : base(name, value)
+    private Rating(string name, byte value) : base(name, value)
     {
     }
 }
